clicking on channel thumbnail also loads newest 5 videos

+++ viewer/routes.py
@@ -23,11 +23,14 @@ def utc_to_local(utc_dt_string):
 def downloadThumbnail(videoID):
     url = f"https://i.ytimg.com/vi/{videoID}/hqdefault.jpg"
     picture_path = os.path.join(app.root_path, 'static/thumbnails', f"{videoID}.jpg")
-    print(f"Downloading image for video {videoID} at {url}")
-    img_data = requests.get(url).content
-    with open(picture_path, 'wb') as handler:
-        handler.write(img_data)
-    print(f"Image for video {videoID} downloaded successfully")
+    if not os.path.exists(picture_path):
+        print(f"Downloading image for video {videoID} at {url}")
+        img_data = requests.get(url).content
+        with open(picture_path, 'wb') as handler:
+            handler.write(img_data)
+        print(f"Image for video {videoID} downloaded successfully")
+    else:
+        print(f"Image for video {videoID} already exists")
     return url_for('static', filename='thumbnails/' + videoID + ".jpg")
 
 def downloadChannelImage(channelID, url):
@@ -46,6 +49,20 @@ def downloadChannelImage(channelID, url):
         print(f"Image for channel {channelID} already exists")
     return url_for('static', filename='thumbnails/' + channelID + ".jpg")
 
+def storeVideoInfoForChannel(channel):
+    videoIDs = channel.videoIDs
+    nVideosAdded = 0
+    for videoID in videoIDs:
+        if not Video.query.filter_by(videoID=videoID).first():
+            v = ytVideo(videoID)
+            thumbnail = downloadThumbnail(videoID)
+            channelImg = downloadChannelImage(v.channelId, channel.img_url)
+            video = Video(title=v.title, channelName=v.channelTitle, channelID=v.channelId, channelImg=channelImg, videoUrl=v.url, videoID=v.videoID, image=thumbnail,description=v.description, publishedAt=utc_to_local(v.publishedAt))
+            db.session.add(video)
+            nVideosAdded += 1
+    db.session.commit()
+    return nVideosAdded
+
 @app.route("/")
 @app.route("/home", methods=['GET', 'POST'])
 def home():
@@ -53,20 +70,8 @@ def home():
     form = SearchForm()
     if form.validate_on_submit():
         channel = channelPlaylist(form.channelName.data, maxResults=form.maxResults.data, isUser=form.isUser.data)
-        print(channel.img_url)
         videoIDs = channel.videoIDs
-        nVideosAdded = 0
-        for videoID in videoIDs:
-            if not Video.query.filter_by(videoID=videoID).first():
-                v = ytVideo(videoID)
-                thumbnail = downloadThumbnail(videoID)
-                channelImg = downloadChannelImage(v.channelId, channel.img_url)
-                video = Video(title=v.title, channelName=v.channelTitle, channelID=v.channelId, channelImg=channelImg, videoUrl=v.url, videoID=v.videoID, image=thumbnail,description=v.description, publishedAt=utc_to_local(v.publishedAt))
-                db.session.add(video)
-                nVideosAdded += 1
-            else:
-                print(Video.query.filter_by(videoID=videoID).first().image)
-        db.session.commit()
+        nVideosAdded = storeVideoInfoForChannel(channel)
         channelID = Video.query.filter_by(videoID=videoIDs[0]).first().channelID
         print(f"ADDED {nVideosAdded} VIDEO(S) TO DATABASE")
         return redirect(url_for('channel', channelID=channelID))
@@ -77,9 +82,18 @@ def home():
 @app.route("/results", methods=['GET', 'POST'])
 def results():
     videos = Video.query.all()
+    videos.sort(key=lambda v: v.publishedAt, reverse=True)
     return render_template('results.html', videos=videos)
 
 @app.route("/channel/<channelID>")
 def channel(channelID):
+
+    channel = channelPlaylist(channelID, maxResults=5, isUser=False)
+    videoIDs = channel.videoIDs
+    nVideosAdded = storeVideoInfoForChannel(channel)
+    channelID = Video.query.filter_by(videoID=videoIDs[0]).first().channelID
+    print(f"ADDED {nVideosAdded} VIDEO(S) TO DATABASE")
+
     videos = Video.query.filter_by(channelID=channelID).all()
+    videos.sort(key=lambda v: v.publishedAt, reverse=True)
     return render_template('results.html', videos=videos)
